removed redundant route state on interface, renamed and relocated interface
